Refactored `FunctionEndpointGenerator` to simplify and optimize type management and parsing logic. Removed unnecessary `FunctionEndpointGeneratorKnownTypes.cs` file.
